refactor(codegen): separate generation responsibilities
