+++ internal/cmd/codegen/generate/webffi/generate.go
@@ -23,7 +23,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"text/template"
 
@@ -53,59 +52,6 @@ var (
 	workerWrapJsFileText string
 )
 
-type managerFuncBodySpec struct {
-	call      string
-	fallback  []string
-	boolAsInt bool
-}
-
-var inputCacheManagerFuncBodies = map[string]managerFuncBodySpec{
-	"GDExtensionSpxInputGetGlobalMousePos": {
-		call: "WebInputMousePos(func() Vec2",
-		fallback: []string{
-			"_retValue := API.SpxInputGetGlobalMousePos.Invoke()",
-			"return JsToGdVec2(_retValue)",
-		},
-	},
-	"GDExtensionSpxInputGetKey": {
-		call: "WebInputKeyState(key, func() bool",
-		fallback: []string{
-			"arg0Low, arg0High := JsSplitGdInt(key)",
-			"_retValue := API.SpxInputGetKey.Invoke(arg0Low, arg0High)",
-			"return JsToGdBool(_retValue)",
-		},
-	},
-	"GDExtensionSpxInputGetMouseState": {
-		call: "WebInputMouseState(mouse_id, func() bool",
-		fallback: []string{
-			"arg0Low, arg0High := JsSplitGdInt(mouse_id)",
-			"_retValue := API.SpxInputGetMouseState.Invoke(arg0Low, arg0High)",
-			"return JsToGdBool(_retValue)",
-		},
-	},
-	"GDExtensionSpxInputGetKeyState": {
-		call: "WebInputKeyState(key, func() bool",
-		fallback: []string{
-			"arg0Low, arg0High := JsSplitGdInt(key)",
-			"_retValue := API.SpxInputGetKeyState.Invoke(arg0Low, arg0High)",
-			"return JsToGdInt(_retValue) != 0",
-		},
-		boolAsInt: true,
-	},
-	"GDExtensionSpxInputGetAxis": {
-		call: "CachedActionAxis(neg_action, pos_action, func() float64",
-		fallback: []string{
-			"arg0 := JsFromGdString(neg_action)",
-			"arg1 := JsFromGdString(pos_action)",
-			"_retValue := API.SpxInputGetAxis.Invoke(arg0, arg1)",
-			"return JsToGdFloat(_retValue)",
-		},
-	},
-	"GDExtensionSpxInputIsActionPressed":      actionBoolManagerFuncBody("pressed", "API.SpxInputIsActionPressed"),
-	"GDExtensionSpxInputIsActionJustPressed":  actionBoolManagerFuncBody("just_pressed", "API.SpxInputIsActionJustPressed"),
-	"GDExtensionSpxInputIsActionJustReleased": actionBoolManagerFuncBody("just_released", "API.SpxInputIsActionJustReleased"),
-}
-
 // Generator renders bindings using metadata owned by one generation task.
 type Generator struct {
 	*common.GenerationContext
@@ -133,18 +79,8 @@ func (g *Generator) Generate(projectPath, spxModulePath string) error {
 
 func GenerateCallbackGoFile(projectPath string, ast clang.CHeaderFileAST) error {
 	funcs := template.FuncMap{
-		"gdiVariableName":       common.GdiVariableName,
-		"snakeCase":             strcase.ToSnake,
-		"camelCase":             strcase.ToCamel,
-		"upper":                 strings.ToUpper,
-		"goReturnType":          common.GoReturnType,
-		"goArgumentType":        common.GoArgumentType,
-		"goEnumValue":           common.GoEnumValue,
 		"add":                   common.Add,
-		"cgoCastArgument":       common.CgoCastArgument,
-		"cgoCastReturnType":     common.CgoCastReturnType,
-		"cgoCleanUpArgument":    common.CgoCleanUpArgument,
-		"trimPrefix":            common.TrimPrefix,
+		"trimPrefix":            strings.TrimPrefix,
 		"mustPrimitiveTypeName": common.MustPrimitiveTypeName,
 	}
 
@@ -154,18 +90,8 @@ func GenerateCallbackGoFile(projectPath string, ast clang.CHeaderFileAST) error 
 
 func GenerateWorkerWrapJsFile(projectPath string, ast clang.CHeaderFileAST) error {
 	funcs := template.FuncMap{
-		"gdiVariableName":    common.GdiVariableName,
-		"snakeCase":          strcase.ToSnake,
-		"camelCase":          strcase.ToCamel,
-		"upper":              strings.ToUpper,
-		"goReturnType":       common.GoReturnType,
-		"goArgumentType":     common.GoArgumentType,
-		"goEnumValue":        common.GoEnumValue,
-		"add":                common.Add,
-		"cgoCastArgument":    common.CgoCastArgument,
-		"cgoCastReturnType":  common.CgoCastReturnType,
-		"cgoCleanUpArgument": common.CgoCleanUpArgument,
-		"trimPrefix":         common.TrimPrefix,
+		"snakeCase":  strcase.ToSnake,
+		"trimPrefix": strings.TrimPrefix,
 	}
 
 	return common.GenerateFile(funcs, "worker.wrap.gen.js", workerWrapJsFileText, ast,
@@ -174,17 +100,7 @@ func GenerateWorkerWrapJsFile(projectPath string, ast clang.CHeaderFileAST) erro
 
 func GenerateGDExtensionInterfaceGoFile(projectPath string, ast clang.CHeaderFileAST) error {
 	funcs := template.FuncMap{
-		"gdiVariableName":     common.GdiVariableName,
-		"snakeCase":           strcase.ToSnake,
-		"camelCase":           strcase.ToCamel,
-		"goReturnType":        common.GoReturnType,
-		"goArgumentType":      common.GoArgumentType,
-		"goEnumValue":         common.GoEnumValue,
-		"add":                 common.Add,
-		"cgoCastArgument":     common.CgoCastArgument,
-		"cgoCastReturnType":   common.CgoCastReturnType,
-		"cgoCleanUpArgument":  common.CgoCleanUpArgument,
-		"trimPrefix":          common.TrimPrefix,
+		"trimPrefix":          strings.TrimPrefix,
 		"loadProcAddressName": common.LoadProcAddressName,
 	}
 
@@ -194,21 +110,10 @@ func GenerateGDExtensionInterfaceGoFile(projectPath string, ast clang.CHeaderFil
 
 func (g *Generator) GenerateManagerWrapperGoFile(projectPath string) error {
 	funcs := template.FuncMap{
-		"gdiVariableName":     common.GdiVariableName,
-		"snakeCase":           strcase.ToSnake,
-		"camelCase":           strcase.ToCamel,
-		"goReturnType":        common.GoReturnType,
-		"goArgumentType":      common.GoArgumentType,
-		"goEnumValue":         common.GoEnumValue,
-		"add":                 common.Add,
-		"cgoCastArgument":     common.CgoCastArgument,
-		"cgoCastReturnType":   common.CgoCastReturnType,
-		"cgoCleanUpArgument":  common.CgoCleanUpArgument,
-		"trimPrefix":          common.TrimPrefix,
-		"isManagerMethod":     g.IsManagerMethod,
-		"getManagerFuncName":  g.ManagerMethodSignature,
-		"getManagerFuncBody":  g.getManagerFuncBody,
-		"getManagerInterface": g.ManagerInterfaceSignature,
+		"camelCase":          strcase.ToCamel,
+		"isManagerMethod":    g.IsManagerMethod,
+		"getManagerFuncName": g.ManagerMethodSignature,
+		"getManagerFuncBody": g.getManagerFuncBody,
 	}
 
 	return common.GenerateFile(funcs, "manager_web.gen.go", managerWebText, g.ManagerData(),
@@ -217,20 +122,10 @@ func (g *Generator) GenerateManagerWrapperGoFile(projectPath string) error {
 
 func (g *Generator) GenerateJsEngineJsFile(projectPath, spxModulePath string) error {
 	funcs := template.FuncMap{
-		"gdiVariableName":     common.GdiVariableName,
-		"snakeCase":           strcase.ToSnake,
-		"camelCase":           strcase.ToCamel,
 		"goReturnType":        common.GoReturnType,
-		"goArgumentType":      common.GoArgumentType,
-		"goEnumValue":         common.GoEnumValue,
-		"add":                 common.Add,
 		"sub":                 common.Sub,
 		"getJsFuncArgs":       g.getJsFuncArgs,
-		"cgoCastArgument":     common.CgoCastArgument,
-		"cgoCastReturnType":   common.CgoCastReturnType,
-		"cgoCleanUpArgument":  common.CgoCleanUpArgument,
 		"getJsFuncBody":       g.getJsFuncBody,
-		"trimPrefix":          common.TrimPrefix,
 		"loadProcAddressName": common.LoadProcAddressName,
 	}
 
@@ -252,326 +147,4 @@ func trimTrailingWhitespace(src []byte) []byte {
 		lines[i] = bytes.TrimRight(line, " \t")
 	}
 	return bytes.Join(lines, []byte("\n"))
-}
-
-func (g *Generator) getManagerFuncBody(function *clang.TypedefFunction) string {
-	if body, ok := getInputCacheManagerFuncBody(function.Name); ok {
-		return body
-	}
-
-	sb := strings.Builder{}
-	prefixTab := "\t"
-	params := []string{}
-	args := common.EffectiveArguments(function)
-	// convert arguments
-	for i, arg := range args {
-		if g.ShouldSkipHighLevelArgument(function, arg) {
-			continue
-		}
-		sb.WriteString(prefixTab)
-		if g.IsNativeArrayDataArg(function, arg) {
-			argName := "arg" + strconv.Itoa(i)
-			sb.WriteString(argName)
-			sb.WriteString(" := JsFromGdArray(")
-			sb.WriteString(g.EffectiveGoArgumentName(function, arg))
-			sb.WriteString(")\n")
-			params = append(params, argName)
-			continue
-		}
-		typeName := common.MustPrimitiveTypeName(arg, function.Name)
-		if g.usesFlatJsGdIntArg(function, arg) {
-			argName := "arg" + strconv.Itoa(i)
-			lowName := argName + "Low"
-			highName := argName + "High"
-			fmt.Fprintf(&sb, "%s, %s := ", lowName, highName)
-			sb.WriteString(flatJsSplitHelper(typeName))
-			sb.WriteString("(")
-			sb.WriteString(g.EffectiveGoArgumentName(function, arg))
-			sb.WriteString(")\n")
-			params = append(params, lowName, highName)
-			continue
-		}
-		argName := "arg" + strconv.Itoa(i)
-		sb.WriteString(argName)
-		sb.WriteString(" := ")
-		sb.WriteString("JsFrom")
-		sb.WriteString(typeName)
-		sb.WriteString("(")
-		sb.WriteString(g.EffectiveGoArgumentName(function, arg))
-		sb.WriteString(")")
-
-		sb.WriteString("\n")
-		params = append(params, argName)
-	}
-
-	// call the function
-	sb.WriteString(prefixTab)
-	if common.HasEffectiveReturn(function) {
-		sb.WriteString("_retValue := ")
-	}
-
-	funcName := "API.Spx" + (common.TrimPrefix(function.Name, "GDExtensionSpx"))
-	sb.WriteString(funcName)
-	sb.WriteString(".Invoke(")
-	for i, param := range params {
-		sb.WriteString(param)
-		if i != len(params)-1 {
-			sb.WriteString(", ")
-		}
-	}
-	sb.WriteString(")")
-
-	if common.HasEffectiveReturn(function) {
-		sb.WriteByte('\n')
-		sb.WriteString(prefixTab)
-		sb.WriteString("return ")
-		typeName := common.EffectiveRawReturnType(function)
-		name := strcase.ToCamel(typeName)
-		if name == "GdObj" {
-			name = "GdObject"
-		}
-		fmt.Fprintf(&sb, "JsTo%s(_retValue)", name)
-	}
-	return sb.String()
-}
-
-func getInputCacheManagerFuncBody(name string) (string, bool) {
-	spec, ok := inputCacheManagerFuncBodies[name]
-	if !ok {
-		return "", false
-	}
-	if spec.boolAsInt {
-		return cachedBoolAsIntBody(spec.call, spec.fallback...), true
-	}
-	return cachedReturnBody(spec.call, spec.fallback...), true
-}
-
-func actionBoolManagerFuncBody(kind, api string) managerFuncBodySpec {
-	return managerFuncBodySpec{
-		call: fmt.Sprintf("CachedActionBool(%q, action, func() bool", kind),
-		fallback: []string{
-			"arg0 := JsFromGdString(action)",
-			"_retValue := " + api + ".Invoke(arg0)",
-			"return JsToGdBool(_retValue)",
-		},
-	}
-}
-
-func cachedReturnBody(call string, fallbackLines ...string) string {
-	lines := []string{"return " + call + " {"}
-	lines = appendIndented(lines, "\t\t", fallbackLines...)
-	lines = append(lines, "\t})")
-	return strings.Join(lines, "\n")
-}
-
-func cachedBoolAsIntBody(call string, fallbackLines ...string) string {
-	lines := []string{"if " + call + " {"}
-	lines = appendIndented(lines, "\t\t", fallbackLines...)
-	lines = append(lines,
-		"\t}) {",
-		"\t\treturn 1",
-		"\t}",
-		"\treturn 0",
-	)
-	return strings.Join(lines, "\n")
-}
-
-func appendIndented(lines []string, indent string, values ...string) []string {
-	for _, value := range values {
-		lines = append(lines, indent+value)
-	}
-	return lines
-}
-
-func (g *Generator) getJsFuncArgs(function *clang.TypedefFunction) []string {
-	args := common.EffectiveArguments(function)
-	result := make([]string, 0, len(args))
-	for _, arg := range args {
-		if g.ShouldSkipHighLevelArgument(function, arg) {
-			continue
-		}
-		argName := g.EffectiveGoArgumentName(function, arg)
-		if g.usesFlatJsGdIntArg(function, arg) {
-			result = append(result, argName+"_low", argName+"_high")
-			continue
-		}
-		result = append(result, argName)
-	}
-	return result
-}
-
-func (g *Generator) usesFlatJsGdIntArg(function *clang.TypedefFunction, arg clang.Argument) bool {
-	if function == nil {
-		return false
-	}
-	if g.IsNativeArrayDataArg(function, arg) {
-		return false
-	}
-	return isFlatJsGdIntLikeType(common.MustPrimitiveTypeName(arg, function.Name))
-}
-
-func isFlatJsGdIntLikeType(typeName string) bool {
-	switch typeName {
-	case "GdInt", "GdObj":
-		return true
-	default:
-		return false
-	}
-}
-
-func flatJsCtor(typeName string) string {
-	switch typeName {
-	case "GdInt":
-		return "Module['_gdspx_new_int']"
-	case "GdObj":
-		return "Module['_gdspx_new_obj']"
-	default:
-		panic(fmt.Sprintf("unsupported flat js gdint-like type: %s", typeName))
-	}
-}
-
-func flatJsSplitHelper(typeName string) string {
-	switch typeName {
-	case "GdInt":
-		return "JsSplitGdInt"
-	case "GdObj":
-		return "JsSplitGdObj"
-	default:
-		panic(fmt.Sprintf("unsupported flat js gdint-like type: %s", typeName))
-	}
-}
-
-func flatJsScratchAccessor(typeName string) string {
-	switch typeName {
-	case "GdInt":
-		return "this._getGdIntScratch()"
-	case "GdObj":
-		return "this._getGdObjScratch()"
-	default:
-		panic(fmt.Sprintf("unsupported flat js gdint-like type: %s", typeName))
-	}
-}
-
-func (g *Generator) getJsFuncBody(function *clang.TypedefFunction) string {
-	if function.Name == "GDExtensionSpxResFreeStr" {
-		// Web strings are values; the wrapper owns its storage.
-		return "// Web strings are value-owned; there is no pointer to release.\n\treturn;"
-	}
-	if spec, ok := g.GetArrayTransformBridgeSpec(function.Name); ok {
-		return "var _fastRetValue = TryArrayTransformFastPath(_gdFuncPtr, " + spec.ArrayArgName + ", " +
-			strconv.Itoa(int(spec.InputArrayType)) + ", " + strconv.Itoa(int(spec.OutputArrayType)) + ", " +
-			strconv.Itoa(spec.OutputCountScale) + ");\n" +
-			"\tif (_fastRetValue == null) {\n" +
-			"\t\tthrow new Error(\"gd" + common.LoadProcAddressName(function.Name) + " fast path unavailable\");\n" +
-			"\t}\n" +
-			"\treturn _fastRetValue"
-	}
-	if function.Name == "GDExtensionSpxInputWriteSnapshot" {
-		return "var _arg0 = RequireWasmFastArray(out, \"gdspx_input_write_snapshot\", 2);\n" +
-			"\tvar _arg1 = FastArrayCount(out);\n" +
-			"\t_gdFuncPtr(_arg0, _arg1);"
-	}
-	if spec, ok := g.GetNativeArrayBridgeSpec(function.Name); ok {
-		if common.HasEffectiveReturn(function) {
-			panic(fmt.Sprintf("native-array webffi path does not support return values: %s", function.Name))
-		}
-		argName := spec.BaseArgName
-		return "var _arg0 = RequireFastArray(" + argName + ", \"gd" + common.LoadProcAddressName(function.Name) + "\", " + strconv.Itoa(int(spec.FastArrayType)) + ");\n" +
-			"\tvar _arg1 = FastArrayCount(" + argName + ");\n" +
-			"\t_gdFuncPtr(_arg0, _arg1);"
-	}
-	if function.Name == "GDExtensionSpxInputGetGlobalMousePos" {
-		return "var _retValue = AllocGdVec2();\n" +
-			"\t_gdFuncPtr(_retValue);\n" +
-			"\tvar _scratch = this._inputMousePosScratch;\n" +
-			"\tvar _floatIndex = _retValue / 4;\n" +
-			"\tvar _heap = Module['HEAPF32'];\n" +
-			"\t_scratch['x'] = _heap[_floatIndex];\n" +
-			"\t_scratch['y'] = _heap[_floatIndex + 1];\n" +
-			"\tFreeGdVec2(_retValue);\n" +
-			"\treturn _scratch"
-	}
-	sb := strings.Builder{}
-	prefixTab := "\t"
-	params := []string{}
-	args := common.EffectiveArguments(function)
-	rawRetType := common.EffectiveRawReturnType(function)
-
-	// call the function
-	if rawRetType != "" {
-		fmt.Fprintf(&sb, "var _retValue = Alloc%s();", rawRetType)
-	}
-	sb.WriteString("\n")
-
-	// convert arguments
-	for i, arg := range args {
-		sb.WriteString(prefixTab)
-		typeName := common.MustPrimitiveTypeName(arg, function.Name)
-		argName := "_arg" + strconv.Itoa(i)
-		if g.usesFlatJsGdIntArg(function, arg) {
-			fmt.Fprintf(&sb, "var %s = ", argName)
-			sb.WriteString(flatJsCtor(typeName))
-			sb.WriteString("(")
-			fmt.Fprintf(&sb, "%s_high, %s_low", arg.Name, arg.Name)
-			sb.WriteString(");")
-
-			sb.WriteString("\n")
-			params = append(params, argName)
-			continue
-		}
-		fmt.Fprintf(&sb, "var %s = ", argName)
-		sb.WriteString("To")
-		sb.WriteString(typeName)
-		sb.WriteString("(")
-		sb.WriteString(arg.Name)
-		sb.WriteString(");")
-
-		sb.WriteString("\n")
-		params = append(params, argName)
-	}
-	sb.WriteString(prefixTab)
-	sb.WriteString("_gdFuncPtr")
-	sb.WriteString("(")
-	for i, param := range params {
-		sb.WriteString(param)
-		if i != len(params)-1 {
-			sb.WriteString(", ")
-		}
-	}
-	if rawRetType != "" {
-		if len(params) > 0 {
-			sb.WriteString(", ")
-		}
-		sb.WriteString("_retValue")
-	}
-	sb.WriteString(");")
-
-	sb.WriteString("\n")
-	// convert arguments
-	for i, arg := range args {
-		sb.WriteString(prefixTab)
-		typeName := common.MustPrimitiveTypeName(arg, function.Name)
-		argName := "_arg" + strconv.Itoa(i)
-		fmt.Fprintf(&sb, "Free%s(%s); \n", typeName, argName)
-	}
-
-	if rawRetType != "" {
-		sb.WriteString(prefixTab)
-		sb.WriteString("var _finalRetValue = ")
-		if isFlatJsGdIntLikeType(rawRetType) {
-			sb.WriteString("this._readGdIntLike(_retValue, ")
-			sb.WriteString(flatJsScratchAccessor(rawRetType))
-			sb.WriteString(");\n")
-		} else {
-			typeName := rawRetType
-			funcName := strcase.ToCamel(typeName)
-			funcName = "ToJs" + strings.ReplaceAll(funcName, "Gd", "")
-			sb.WriteString(funcName)
-			sb.WriteString("(_retValue);\n")
-		}
-		fmt.Fprintf(&sb, "%sFree%s(_retValue); \n", prefixTab, rawRetType)
-		sb.WriteString(prefixTab)
-		sb.WriteString("return _finalRetValue")
-	}
-	return sb.String()
 }
